requirements.txt
Flask==2.1.0
requests
Werkzeug==2.1.2
joblib==1.2.0
pandas==1.3.5
numpy==1.21.6
matplotlib==3.2.2
scikit-learn==1.0.2
tensorflow==2.9.2
Pillow==8.4.0
opencv-python==4.1.2.30
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
seaborn==0.11.1
PyYAML==6.0
tensorboard==2.9.1
scipy==1.9.3
seaborn==0.11.1


